\name{DataExp}
\alias{DataExp}
\docType{data}
\title{DataExp}
\description{A dataset in the form of stacked time-series. Quarterly data cover nine countries (Poland, Bulgaria, the Czech Republic, Hungary, Latvia, Lithuania, Romania, Slovakia, Slovenia; the numbers denoting particular countries form a series i=1,2,...,9) from 2005q2 to 2013q4. The dataset contains the following (below) variables}
\usage{data("DataExp")}
\format{
  A data frame with 315 observations on the following 16 variables.
  \describe{
    \item{\code{y10}}{a numeric vector}
    \item{\code{y10spread}}{a numeric vector}
    \item{\code{riskavers}}{a numeric vector}
    \item{\code{debt}}{a numeric vector}
    \item{\code{deficit}}{a numeric vector}
    \item{\code{openess}}{a numeric vector}
    \item{\code{cpi}}{a numeric vector}
    \item{\code{growth}}{a numeric vector}
    \item{\code{crisk}}{a numeric vector}
    \item{\code{urate}}{a numeric vector}
    \item{\code{iip}}{a numeric vector}
    \item{\code{iipnetto}}{a numeric vector}
    \item{\code{cagdp}}{a numeric vector}
    \item{\code{caresvs}}{a numeric vector}
    \item{\code{cds}}{a numeric vector}
    \item{\code{bidask}}{a numeric vector}
  }
}
\examples{
data(DataExp)
}
\keyword{panel data}
