\name{PooledMeanGroup-package}
\alias{PooledMeanGroup-package}
\alias{PooledMeanGroup}
\docType{package}
\title{
\packageTitle{PooledMeanGroup}
}
\description{Calculates the pool mean group (PMG) estimator for dynamic panel data models, as described by Pesaran, Shin and Smith (1999) <doi:10.1080/01621459.1999.10474156>. This estimator enables the intercepts, short-run coefficient and error variances to differ freely across groups, but restricts the long-run coefficients to being equal. Additionally, it allows the numbers of time series observations to differ freely across groups. This software also performs diagnostic tests of error terms, such as autocorrelation, heteroscedasticity and normality.
\packageDescription{PooledMeanGroup}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PooledMeanGroup}
\packageIndices{PooledMeanGroup}
}
\author{Lech Kujawski, Piotr Zientara
\packageAuthor{PooledMeanGroup}

Maintainer: \packageMaintainer{PooledMeanGroup}
}
\references{
Pesaran, Shin and Smith (1999) <doi:10.1080/01621459.1999.10474156>
}
\keyword{package, PMG, pooled mean group, dynamic panel models}


